Created django project, base app and skeleton objs.
Set up django project with a complete readme for quick setup. Set up
very basic models, serializers, and views to be used as examples for the
full versions of each. Updated initial settings to work with mysql,
verified that mysql works and tested to make sure the basic urls load
when the server is run.